Flash attention for GPUs like in maxtext (#149)
* add support for flux vae. ~ wip

* test for flux vae both encoding and decoding.

* add clip text encoder test

* remove transformers inside maxdiffusion, add transformers dependency. Start creating generation code for flux.

* add double block to flux

* forward pass for single double block.

* trying to use scan.

* add single stream block

* finish transformer

* convert pt weights to flax and load transformer state.

* apply fsdp sharding, do one forward pass in the transformer.

* wip - generate fn

* working loop, bad generation

* e2e, encoder offloading.

* support both dev and schnell loading. Images still incorrect.

* flux schnell working

* removed unused code.

* support dev

* add sentencepiece requirement

* fix repeated double and single blocks.

* optimized flash block sizes for trillium.

* clean up code and lint

* fix sdxl generate smoke tests.

* fix rest of unit tests.

* update readme and some dependencies.

* remove unused dependencies.

* initial lora implementation for flux

* adding another format lora support.

* Support other format loras. update readme. Run code_style.

* ruff

* fix typo in readme.

* Added FA support for GPUs

* ruff and code_style

* fixed final comments

* Correcting small misstake due to missunderstanding

---------

+++ src/maxdiffusion/generate_flux.py
@@ -77,7 +77,7 @@ def unpack(x: Array, height: int, width: int) -> Array:
 
 
 def vae_decode(latents, vae, state, config):
-  img = unpack(x=latents, height=config.resolution, width=config.resolution)
+  img = unpack(x=latents.astype(jnp.float32), height=config.resolution, width=config.resolution)
   img = img / vae.config.scaling_factor + vae.config.shift_factor
   img = vae.apply({"params": state.params}, img, deterministic=True, method=vae.decode).sample
   return img
